refactor(kernel): consolidate device probing into a driver registry
Replace direct calls to `sd_card::init()` and `usb_storage::init()` with a unified
`DriverRegistry` that scans PCI devices and matches drivers. The new `device_probe`
init step builds the registry and probes all buses, while the old `usb_storage` and
`sd_card` steps become serial-log markers only. GUI and shell hooks now use the registry
global state (`registry::USB_DRIVES`, `registry::SD_DRIVES`, etc.) instead of module-specific
statics. The `sd_card` and `usb_storage` modules are removed from `mod.rs` and the `registry`
module is added.
