+++ fullerene-kernel/src/gui.rs
@@ -130,13 +130,13 @@ pub fn init() {
             result
         }),
         usb_drive_list: Some(|| {
-            let drives = crate::drivers::usb_storage::USB_DRIVES.lock();
+            let drives = crate::drivers::registry::USB_DRIVES.lock();
             drives
                 .iter()
                 .map(|d| (d.name.clone(), d.mount_point.clone()))
                 .collect()
         }),
-        usb_poll: Some(|| crate::drivers::usb_storage::poll_usb()),
+        usb_poll: Some(|| crate::drivers::registry::poll_usb()),
         shell_cmd: None,
         launch_shell: Some(|| {
             crate::scheduler::request_shell_launch();
